requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.30.6
yt-dlp==2026.3.17
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
